pyproject.toml
[tool.poetry]
name = "this-viz-will-kill-you"
version = "0.1.0"
description = ""
authors = ["Nicole Pearce"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
pandas = "^2.1.2"
numpy = "^1.26.1"
pyyaml = "^6.0.1"
loguru = "^0.7.2"
pytest = "^7.4.3"
pytest-mock = "^3.12.0"
click = "^8.1.7"
spotipy = "^2.23.0"
pyarrow = "^13.0.0"
spacy = {extras = ["apple"], version = "^3.7.2"}
gensim = "^4.3.2"
seaborn = "^0.13.0"
python-docx = "^1.0.1"
nltk = "^3.8.1"
python-graphql-client = "^0.4.3"


[tool.poetry.group.dev.dependencies]
mypy = "^1.6.1"
ipykernel = "^6.26.0"

[tool.pytest.ini_options]
testpaths = "./tests"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
